Move HIV status methods/attributes into own module
